# tmp/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GROWI REST API v3",
    "version": "4.3.0-RC"
  },
  "servers": [
    {
      "url": "https://demo.growi.org{basePath}",
      "variables": {
        "basePath": {
          "default": "/_api/v3",
          "description": "base path"
        }
      }
    }
  ],
  "security": [
    {
      "api_key": []
    }
  ],
  "components": {
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "access_token",
        "in": "query"
      }
    },
    "schemas": {
...
      "PersonalSettings": {
        "description": "personal settings",
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "lang": {
            "type": "string"
          },
          "isEmailPublished": {
            "type": "boolean"
          }
        }
      },
      "Passwords": {
        "description": "passwords for update",
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          },
          "newPasswordConfirm": {
            "type": "string"
          }
        }
      },
      "AssociateUser": {
        "description": "Ldap account for associate",
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "DisassociateUser": {
        "description": "Ldap account for disassociate",
        "type": "object",
        "properties": {
          "providerType": {
            "type": "string"
          },
          "accountId": {
            "type": "string"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/personal-setting": {
      "get": {
        "tags": [
          "PersonalSetting"
        ],
        "operationId": "getPersonalSetting",
        "summary": "/personal-setting",
        "description": "Get personal parameters",
        "responses": {
          "200": {
            "description": "params of personal",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "currentUser": {
                      "type": "object",
                      "description": "personal params"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "PersonalSetting"
        ],
        "operationId": "updatePersonalSetting",
        "summary": "/personal-setting",
        "description": "Update personal setting",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonalSettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "params of personal",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "currentUser": {
                      "type": "object",
                      "description": "personal params"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/personal-setting/password": {
      "put": {
        "tags": [
          "PersonalSetting"
        ],
        "operationId": "putUserPassword",
        "summary": "/personal-setting/password",
        "description": "Update user password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Passwords"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "user password",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "userData": {
                      "type": "object",
                      "description": "user data updated"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/personal-setting/associate-ldap": {
      "put": {
        "tags": [
          "PersonalSetting"
        ],
        "operationId": "associateLdapAccount",
        "summary": "/personal-setting/associate-ldap",
        "description": "associate Ldap account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssociateUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "succeded to associate Ldap account",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "associateUser": {
                      "type": "object",
                      "description": "Ldap account associate to me"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/personal-setting/disassociate-ldap": {
      "put": {
        "tags": [
          "PersonalSetting"
        ],
        "operationId": "disassociateLdapAccount",
        "summary": "/personal-setting/disassociate-ldap",
        "description": "disassociate Ldap account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisassociateUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "succeded to disassociate Ldap account",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "disassociateUser": {
                      "type": "object",
                      "description": "Ldap account disassociate to me"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/personal-setting/associate-mikan": {
      "put": {
        "tags": [
          "PersonalSetting"
        ],
        "operationId": "associateMikanAccount",
        "summary": "/personal-setting/associate-mikan",
        "description": "associate Mikan account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssociateUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "succeded to associate Mikan account",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "associateUser": {
                      "type": "object",
                      "description": "Mikan account associate to me"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/personal-setting/disassociate-mikan": {
      "put": {
        "tags": [
          "PersonalSetting"
        ],
        "operationId": "disassociateMikanAccount",
        "summary": "/personal-setting/disassociate-mikan",
        "description": "disassociate Mikan account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisassociateUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "succeded to disassociate Mikan account",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "disassociateUser": {
                      "type": "object",
                      "description": "Mikan account disassociate to me"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}